feat(activity): refresh on org switch and add member/guest views
Org-scope the feed reset key and the filter option-list loaders so a switch resets to page 1 and refetches, fixing stale rows and the Export CSV button acting on the previous org's view.

Replace the admin-only redirect with role-based content: owners/admins get the real feed, members get a labelled sample (mock data, no audit request), and signed-out users get an in-page sign-in. Show the nav item to everyone and route it while signed-out.

Extract the shared activity row into its own module so the feed and the member sample render identically.

Refs KEEP-842
